Split pr_cmd_inline repo_helpers tests into modules
